# expression: implement vectorized evaluation for `builtinToDaysSig`